Refactor: Title sequences and switch to size_t

+++ src/openrct2/title/TitleSequenceManager.h
@@ -26,7 +26,7 @@ struct TitleSequenceManagerItem
 {
     std::string Name;
     std::string Path;
-    uint16      PredefinedIndex;
+    size_t      PredefinedIndex;
     bool        IsZip;
 };
 
@@ -41,7 +41,7 @@ namespace TitleSequenceManager
     void Scan();
 }
 
-constexpr uint16 PREDEFINED_INDEX_CUSTOM = UINT16_MAX;
+constexpr size_t PREDEFINED_INDEX_CUSTOM = SIZE_MAX;
 
 extern "C" {
 #endif
@@ -50,7 +50,7 @@ extern "C" {
     const utf8 * title_sequence_manager_get_name(size_t index);
     const utf8 * title_sequence_manager_get_path(size_t index);
     const utf8 * title_sequence_manager_get_config_id(size_t index);
-    uint16 title_sequence_manager_get_predefined_index(size_t index);
+    size_t title_sequence_manager_get_predefined_index(size_t index);
     size_t title_sequence_manager_get_index_for_config_id(const utf8 * configId);
     size_t title_sequence_manager_get_index_for_name(const utf8 * name);
     bool title_sequence_manager_is_name_reserved(const utf8 * name);
